Clang now compiling

+++ libs/ec/inc/public/tfc/ec/devices/beckhoff/EL1xxx.hpp
@@ -22,7 +22,7 @@ public:
 
   void process_data(std::span<std::byte> input, std::span<std::byte>) noexcept final {
     static_assert(size <= 8);
-    std::bitset<size> const in_bits(input[0]);
+    std::bitset<size> const in_bits(static_cast<uint8_t>(input[0]));
     for (size_t i = 0; i < size; i++) {
       bool const value = in_bits.test(i);
       if (value != last_values_[i]) {
